refactor: split tools into single files
