+++ .gitignore
@@ -2,3 +2,8 @@
 .opencode*
 OpenCode.md
 dist/
+CLAUDE.md
+.claude/
+
+# go build artifact
+mcp-filesystem-server
