+++ filesystemserver/handler/create_directory.go
@@ -1,0 +1,112 @@
+package handler
+
+import (
+	"context"
+	"fmt"
+	"os"
+
+	"github.com/mark3labs/mcp-go/mcp"
+)
+
+func (fs *FilesystemHandler) HandleCreateDirectory(
+	ctx context.Context,
+	request mcp.CallToolRequest,
+) (*mcp.CallToolResult, error) {
+	path, err := request.RequireString("path")
+	if err != nil {
+		return nil, err
+	}
+
+	// Handle empty or relative paths like "." or "./" by converting to absolute path
+	if path == "." || path == "./" {
+		// Get current working directory
+		cwd, err := os.Getwd()
+		if err != nil {
+			return &mcp.CallToolResult{
+				Content: []mcp.Content{
+					mcp.TextContent{
+						Type: "text",
+						Text: fmt.Sprintf("Error resolving current directory: %v", err),
+					},
+				},
+				IsError: true,
+			}, nil
+		}
+		path = cwd
+	}
+
+	validPath, err := fs.validatePath(path)
+	if err != nil {
+		return &mcp.CallToolResult{
+			Content: []mcp.Content{
+				mcp.TextContent{
+					Type: "text",
+					Text: fmt.Sprintf("Error: %v", err),
+				},
+			},
+			IsError: true,
+		}, nil
+	}
+
+	// Check if path already exists
+	if info, err := os.Stat(validPath); err == nil {
+		if info.IsDir() {
+			resourceURI := pathToResourceURI(validPath)
+			return &mcp.CallToolResult{
+				Content: []mcp.Content{
+					mcp.TextContent{
+						Type: "text",
+						Text: fmt.Sprintf("Directory already exists: %s", path),
+					},
+					mcp.EmbeddedResource{
+						Type: "resource",
+						Resource: mcp.TextResourceContents{
+							URI:      resourceURI,
+							MIMEType: "text/plain",
+							Text:     fmt.Sprintf("Directory: %s", validPath),
+						},
+					},
+				},
+			}, nil
+		}
+		return &mcp.CallToolResult{
+			Content: []mcp.Content{
+				mcp.TextContent{
+					Type: "text",
+					Text: fmt.Sprintf("Error: Path exists but is not a directory: %s", path),
+				},
+			},
+			IsError: true,
+		}, nil
+	}
+
+	if err := os.MkdirAll(validPath, 0755); err != nil {
+		return &mcp.CallToolResult{
+			Content: []mcp.Content{
+				mcp.TextContent{
+					Type: "text",
+					Text: fmt.Sprintf("Error creating directory: %v", err),
+				},
+			},
+			IsError: true,
+		}, nil
+	}
+
+	resourceURI := pathToResourceURI(validPath)
+	return &mcp.CallToolResult{
+		Content: []mcp.Content{
+			mcp.TextContent{
+				Type: "text",
+				Text: fmt.Sprintf("Successfully created directory %s", path),
+			},
+			mcp.EmbeddedResource{
+				Type: "resource",
+				Resource: mcp.TextResourceContents{
+					URI:      resourceURI,
+					MIMEType: "text/plain",
+					Text:     fmt.Sprintf("Directory: %s", validPath),
+				},
+			},
+		},
+	}, nil
+}
